flask_server/requirements.txt
Flask
Flask-Cors
python-dotenv
pymongo
PyJWT
python-dateutil
google-generativeai
